requirements.txt
torchvision>=0.19.0
opencv-python>=4.9.0.80
diffusers>=0.31.0
transformers>=4.49.0
tokenizers>=0.20.3
accelerate>=1.1.1
tqdm
imageio
easydict
ftfy
dashscope
imageio-ffmpeg
numpy>=1.23.5,<2
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
